mrc-leaf-helpers(move-resolve-language-model,mechanical_provable): Move resolve_language_model from model_runner.py to model_loader/utils.py (cut+paste)
